+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Unreleased
 * Added net routine for extracting subnetwork containing a specific set of buses.
 * Extended problem.show() to show number of vars and constraints of each type.
 * Fix bug involving problem/constraint/function's network going out of scope in Python.
+* Added load reactive power limits. 
 
 Version 1.3.2
 -------------